+++ src/app/modules/select-records-page/select-records-page.component.spec.ts
@@ -32,6 +32,7 @@ import { CartService } from '../../shared/cart/cart.service';
 import {
   SearchParametersComponent
 } from '../search-parameters/search-parameters.component';
+import { ResourceTableComponent } from '../resource-table/resource-table.component';
 
 describe('SelectRecordsPageComponent (when there are studies for the user)', () => {
   let component: SelectRecordsPageComponent;
@@ -85,6 +86,13 @@ describe('SelectRecordsPageComponent (when there are studies for the user)', () 
     loader = TestbedHarnessEnvironment.loader(fixture);
     component = fixture.componentInstance;
     fixture.detectChanges();
+    // The first call to getHarness in the expectNumberOfRecords function will
+    // never end if the interval is active. This looks like a bug.
+    // As a workaround, just do not create the "interval()":
+    spyOn(
+      ResourceTableComponent.prototype,
+      'runPreloadEvents'
+    ).and.callFake(() => {});
   });
 
   afterEach(() => {
@@ -140,8 +148,11 @@ describe('SelectRecordsPageComponent (when there are studies for the user)', () 
    * @param query - value of the q parameter for the request to the CTSS.
    */
   async function loadVariables(query = ''): Promise<void> {
+    (ResourceTableComponent.prototype
+      .runPreloadEvents as jasmine.Spy).calls.reset();
     await selectTab('Variables');
-
+    fixture.detectChanges();
+    expect(component.variableTable.runPreloadEvents).not.toHaveBeenCalled();
     mockHttp
       .expectOne((req: HttpRequest<any>) => {
         return (
@@ -152,6 +163,8 @@ describe('SelectRecordsPageComponent (when there are studies for the user)', () 
         );
       })
       .flush(query ? threeVariables : fourVariables);
+    fixture.detectChanges();
+    expect(component.variableTable.runPreloadEvents).toHaveBeenCalledOnceWith();
     await expectNumberOfRecords(query ? 3 : 4);
   }
 
@@ -466,6 +479,13 @@ describe('SelectRecordsPageComponent (when there are no studies for the user)', 
     fixture.detectChanges();
     await fixture.whenStable();
     fixture.detectChanges();
+    // The first call to getHarness in the expectNumberOfRecords function will
+    // never end if the interval is active. This looks like a bug.
+    // As a workaround, just do not create the "interval()":
+    spyOn(
+      ResourceTableComponent.prototype,
+      'runPreloadEvents'
+    ).and.callFake(() => {});
   });
 
   afterEach(() => {
@@ -494,6 +514,7 @@ describe('SelectRecordsPageComponent (when there are no studies for the user)', 
       .expectOne('$fhir/Observation?_elements=code,value,category&_count=50')
       .flush(observations);
     fixture.detectChanges();
+    expect(component.variableTable.runPreloadEvents).toHaveBeenCalledOnceWith();
     await fixture.whenStable();
     fixture.detectChanges();
     mockHttp